LayoutTests/fast/dom/named-items-with-empty-name.html

<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id=""></div>
    
<script>
description('This tests empty strings do not return a value in collections or named items cases.');

shouldBeTrue("document.querySelectorAll('div')[''] === undefined");
shouldBeTrue("document.getElementsByTagName('div')[''] === undefined");
shouldBeTrue("document.body.children[''] === undefined");
shouldBeTrue("document.all[''] === undefined");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
